umh: Kill process using the proper SIGKILL signal.
When we detect an invalid binary, we forbid it execution by changing the
subprocess execution path with slashes.
On some kernels, this may not work because we have to map the memory page as
writable. At this moment of execution, we can stop executing the process using
the froce_sig to send signals. The send_sig_info will not arrive at the right
time.

Signed-off-by: Mariusz Zaborski <oshogbo@vexillium.org>