Build LKRG in-tree using copy-builtin.sh
DKMS is not always desirable on a system which one intends to
harden from local attackers. The build-chain creates a wide local
attack surface and can be used to build kmode rootkits and other
fun toys for red team. There are also space, performance,
and security considerations such as leaving RANDSTRUCT seeds
around in headers which can be addressed by bringing the module
in with the kernel package.

In order to avoid this, and take advantage of pending advances
in LTO, KCFI, and other methods depending on or benefitting from
having all of the code available at initial compile/link-time, a
script has been added to update the desired kernel tree with the
LKRG source code and relevant Makefile and Kconfig directives.

Since LKRG is still in active development, this script pulls all
relevant content and the current commit hash to let the kernel
maintainers using it keep aware of how far they've fallen behind
and see if subsequent commits merit a rebuild of the in-tree code.

Testing:
 Build-tested against linux-hardened 5.4.84.a as module and linked
directly into the kernel.
 Runtime testing with both build scenarios.