In-tree: fix for 58b9ec06
After 58b9ec06, the script which copies module files into the
kernel source tree does not properly capture the files to be built
into p_lkrg.o.
Fix the awk expression performing the capture to adjust for the
new $(TARGET)-obj syntax and sed in-place to restore the original
name for in-tree build.