p_cmp_creds considers sudo to be a privesc attack
Check whether the current process is "sudo" before running all of
the UID checks.

The stack traces produced priorto this commit show failures in the
EUID, SUID, and FSUID checks.

This may not be the best way to handle this as attackers running
privesc binaries just need to name them "sudo," but it does fix
the runtime issue of trying to use the legitimate sudo bin. A path
check for "/sbin/sudo" might be better, but distributions place
binaries in different places... Ideally we'd have some better way
to determine if the escalating binary is legitimate along with
coverage for other potentially impacted valid privesc methods.