Copy-builtin: optionally use git via LKRG_USE_GIT
As pointed out by Solar Designer, git is somewhat non-germaine to
the script as it's really part of an external workflow replicated
across out-of-tree repos used by Semper Victus for the trees which
we maintain in-house. The meat of the function in ths script puts
data into the right places within a kernel tree, with the commit
semantics being an ancillary function.

Make all of the git operations optional by enclosing them in tests
for the existence of the LKRG_USE_GIT variable.