README: Fixed punctuation, word choice, and word flow
Made a few minor changes, particularly to punctuation, but also to word
choice and word flow. For example, "might or might not" is more
uncomfortable to say and read than "may or may not" as the ending "t"
requires a stop. I believe these changes **may** (heh) increase readability
marginally. I also did not go through and check the commands and their
descriptions.