CI: mkosi-mainline: Support for 'noble' mainline kernels
Ubuntu started to create mainline kernel packages for upcoming 'noble'
release but jammy's (where CI runs) debootstrap knows nothing about it.

debootstrap fails with (though this is not visible in mkosi output):

    E: No such script: /usr/share/debootstrap/scripts/noble

Add Ubuntu's development debootstrap from their git. It's possible to
run it from the git repo itself (as stated in their README). The repo
cloned into `/debootstrap` so we don't need to set `DEBOOTSTRAP_DIR` and
executable is symlinked to `/usr/local/bin` which we prepared earlier
(and GA prepends it to `PATH`).

For the current bug it's not required to symlink `debootstrap` script
because today we only need to update their `scripts` directory but this
fuller update may be future proof.

Fixes: https://github.com/lkrg-org/lkrg/issues/303
Signed-off-by: Vitaly Chikunov <vt@altlinux.org>