CI: Fix alt:sisyphus build
ALT switched to numerical kernel flavours, thus no 'un-def' anymore.
Bring the new package 'kernel-headers-modules-mainline' that's supposed
to track the mainline kernels. If it fails bring 'latest1' kernel
headers. Bringing 'mainline' could fail during the merge window when
`x.y.1` stable is already released to the `x.y` flavour, so the
pseudo-flavour isn't longer 'mainline' but 'latest1' (note it ends with
`1`). This should not be the 'latest' as it provided by multiple
packages and naïve apt-get install will refuse to install.

Fixes: https://github.com/lkrg-org/lkrg/issues/360
Signed-off-by: Vitaly Chikunov <vt@altlinux.org>