CI: Fix and move centos:7 to docker-cross
Native GA container cannot handle `centos:7` anymore, and workarounds are
ugly. Just use native docker, which we have in cross workflow. (Even
though it will not be a cross build. Maybe it will be better renamed do
docker build.)

To support this `cross.yml` is reworked to run any command instead of
getting a simple list of packages from matrix.

Fixes: https://github.com/lkrg-org/lkrg/issues/359
Signed-off-by: Vitaly Chikunov <vt@altlinux.org>