Skip __queue_work() probe when called from hard IRQ context
It doesn't make sense to try and validate the current task when called from
IRQ context because the IRQ handler is the one calling __queue_work(), not
the current task that happened to be interrupted.

Skip the __queue_work() probe when we know the call doesn't come from the
current task.

Signed-off-by: Sultan Alsawaf <sultan@ciq.com>