Fix seccomp races without using an additional kprobe
Unfortunately, the elegant solution of probing the get_seccomp_filter()
call inside seccomp_sync_threads() isn't reliable because that specific
get_seccomp_filter() call site may be inlined as a copy. While it is
possible to alternatively probe {__}put_seccomp_filter(), there is still a
similar degree of risk of copy inlining.

So, scrap the elegant approach for a clunkier one that always works.

Signed-off-by: Sultan Alsawaf <sultan@ciq.com>