Support new kernels without hook-able {override,revert}_creds()
by not validating the effective credentials when the pointer differs, which
may indicate that override_creds() is in effect or that the pointer was
overwritten in an attack.  So with this workaround we'll be blind to cred
pointer overwrite attacks, but would still detect overwrites of the creds
themselves if the pointer is kept intact.  We should also still detect
exploits' use of commit_creds() as that function also replaces real_cred,
which we continue to validate in all cases.

Works around #371