Hooking: Switch many hooks to simple kprobes
For almost all of the hooks changed by this commit, the return hooks
were dummy - just a "return 0;" - but there was one special case:

Our hook of do_exit() previously did a p_ed_enforce_validation() on return.
This didn't make sense since do_exit() isn't supposed to return and since
we'd have removed our shadow entry for the current task by that point.
This is getting removed here along with the switch to simple kprobe.