UMH: Don't set "off" in our hook of call_usermodehelper_exec_async()
This was unneeded because:

1. We're running in a kthread context here, and we do not currently track
   kthreads, so lookup of the shadow task entry for the current task should
   always fail here (and if it doesn't, that's a bug).

2. call_usermodehelper_exec_async() itself only updates the capability
   sets, which we don't yet track, so even if we tracked the task and
   later validated it, we'd currently not have noticed these differences.

We may reconsider (and reintroduce similar code) if we later make changes
such that both of the above conditions become false.

Until we very recently switched the do_exit() hook from a kretprobe to a
simple kprobe, that hook could occasionally fail.  This meant that the
lookup here could occasionally succeed if the kthread happened to reuse the
task_struct address of an exited task that we failed to remove from our
tracking.  Further, the same would happen for a subsequent lookup on exec.
This is why we needed to have this code and the recent workaround allowing
"off" to be set twice, but we should not need that anymore.

Along with the mentioned workaround (now removed here) and the do_exit()
hooking change, this commit completes fixing #385.