Makefile: Add install-module target
This is useful for CI where we need to install `lkrg.ko`, but do not
need to touch init system. Additionally, fix a (seemingly redundant)
depmod invocation, if used with `KERNELRELEASE` pass it to the depmod
call, and while we at it, remove `-a` because it's a default. Make
`install` target dependent on `all` to allow parallel '-j all install`
builds.

Signed-off-by: Vitaly Chikunov <vt@altlinux.org>