Fix JUMP_LABEL batch indexing in hook of arch_jump_label_transform_apply
There were two sub-issues:
1. The index to byte offset multiplication was wrong (multiplied by
   struct size in bytes and then implicitly again by pointer size).
2. The input array was indexed by the output array's index, so it would
   not advance to next index if an element was skipped from output.

Surprisingly, these issues appeared benign so far (detected in code review
only).  We had out-of-bounds reads and failed to process most elements, but
literally nothing happened as a result in my testing.  This may be because
the batch always corresponds to just one module, so it was sufficient to
get the first element right (which we did despite of these bugs) to get
that module's hash recomputed.

Fixes #399