pINT: No longer store/validate override creds
This made little sense because we're not supposed to reach validation when
override is in effect on pre-6.13, and we infer whether override is in
effect from cred != real_cred on 6.13+ (except where we know it can't be,
where we instead insist on cred == real_cred).

This also side-steps the issue where we could access over a stale copy of
the cred pointer.

Fixes #374