Drop section .p_lkrg_read_only
because it was never necessary to begin with: the struct will always have
trailing padding out to its alignment boundary due to stride.  And the
struct's alignment is one page because the strictest alignment requirement
of any of the struct's members is one page.  In other words, our goal of
having no unrelated symbols placed in the same page(s) where our read-only
data is located is achieved anyway.

Fixes #158