pINT: Drop capset-related hooks
These functions may call commit_creds(), but only modifying securebits and
capabilities, which we unfortunately don't validate and probably don't have
code to fully track yet anyway.  We may revert this commit or reimplement
this once we proceed to implement such tracking and validation for real.

See #188

This is a step towards fixing #409